requirements.txt
alembic
bleach
blinker
certifi
charset-normalizer
click
colorama
DateTime
dnspython
dominate
email-validator
Flask
Flask-Bootstrap
Flask-FontAwesome
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
flask-talisman
Flask-WTF
flask
greenlet
gunicorn
idna
itsdangerous
Jinja2
Mako
MarkupSafe
packaging
psycopg2
pybase64
PyJWT
python-dateutil
pytz
requests
setuptools
six
SQLAlchemy
twilio
typing_extensions
urllib3
visitor
webencodings
Werkzeug
WTForms
zope.interface
alembic
bidict
bleach
blinker
cachetools
certifi
charset-normalizer
click
colorama
DateTime
dnspython
dominate
email-validator
Flask-Bootstrap
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-PageDown
Flask-SocketIO
Flask-SQLAlchemy
Flask-WTF
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
Google-Images-Search
googleapis-common-protos
gunicorn
httplib2
idna
itsdangerous
Jinja2
Mako
Markdown
MarkupSafe
packaging
protobuf
psycopg
pyasn1
pyasn1-modules
pybase64
pyfiglet
pyparsing
python-dotenv
python-engineio
python-resize-image
python-socketio
pytz
requests
rsa
six
SQLAlchemy
termcolor
threaded
tzdata
uritemplate
urllib3
visitor
webencodings
Werkzeug
WTForms
zope.interface
faker
fpdf
ffmpeg-python
flask-session
flask-caching
redis
flask-compress
flask_dance
flask_cors
feedparser
imageio_ffmpeg
celery
celery[redis]